requirements.txt
Django
djangorestframework
mysqlclient
uwsgi
drf-yasg
django-filter
